+++ CHANGES.md
@@ -13,6 +13,7 @@ Other changes:
 Bugfix 🐛:
  - Scroll breadcrumbs to top when opened
  - Render default room name when it starts with an emoji (#477)
+ - Do not display " (IRC)") in display names https://github.com/vector-im/riot-android/issues/444
 
 Translations 🗣:
  -